requirements.txt
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
